+++ lib/components/features/event-manage/EventRegistration.tsx
@@ -3,6 +3,7 @@ import { LimitsAndRestrictions } from "./registration/LimitsAndRestrictions";
 import { RegistrationForm } from "./registration-form/RegistrationForm";
 import { useEvent } from "./store";
 import { TicketList } from "./registration/TicketList";
+import { CouponList } from "./registration/CouponList";
 
 export function EventRegistration() {
   const event = useEvent();
@@ -13,6 +14,8 @@ export function EventRegistration() {
     <div className="space-y-8">
       <LimitsAndRestrictions event={event} />
       <TicketList event={event} />
+      <hr className="border-t border-t-divider" />
+      <CouponList event={event} />
       <hr className="border-t border-t-divider" />
       <RegistrationForm />
     </div>